+++ src/main/java/cz/uhk/pro2/tulipani/web/dto/AddUserToTodolistRequest.java
@@ -2,6 +2,6 @@ package cz.uhk.pro2.tulipani.web.dto;
 
 import jakarta.validation.constraints.NotNull;
 
-public record AddUserToTodolistRequest(@NotNull Long userId, Long roleId) {
+public record AddUserToTodolistRequest(@NotNull Integer userId, Integer roleId) {
 }
