+++ src/main/java/cz/uhk/pro2/tulipani/web/dto/AssignUserToTaskRequest.java
@@ -2,6 +2,6 @@ package cz.uhk.pro2.tulipani.web.dto;
 
 import jakarta.validation.constraints.NotNull;
 
-public record AssignUserToTaskRequest(@NotNull Long userId) {
+public record AssignUserToTaskRequest(@NotNull Integer userId) {
 }
