+++ src/main/java/cz/uhk/pro2/tulipani/web/dto/TodolistUserResponse.java
@@ -1,5 +1,5 @@
 package cz.uhk.pro2.tulipani.web.dto;
 
-public record TodolistUserResponse(Long userId, Long roleId, Boolean isListCreator) {
+public record TodolistUserResponse(Integer userId, Integer roleId, Boolean isListCreator) {
 }
