+++ src/test/java/cz/uhk/pro2/tulipani/web/controller/TaskControllerTest.java
@@ -32,8 +32,8 @@ class TaskControllerTest {
 
     @Test
     void createTask_returnsCreatedTask() {
-        CreateTaskRequest request = new CreateTaskRequest("Task A", "Description", null, 1L, 2L);
-        TaskResponse response = new TaskResponse(10L, "Task A", "Description", null, "todo", 1L, 2L, 100L, "100");
+        CreateTaskRequest request = new CreateTaskRequest("Task A", "Description", null, (int) 1L, (int) 2L);
+        TaskResponse response = new TaskResponse((int) 10L, "Task A", "Description", null, "todo", (int) 1L, (int) 2L, (int) 100L, "100");
         when(taskService.createTask(request, "auth-1")).thenReturn(response);
 
         ResponseEntity<TaskResponse> result = taskController.createTask(request, "auth-1");
@@ -45,10 +45,10 @@ class TaskControllerTest {
     @Test
     void updateTaskStatus_returnsUpdatedTask() {
         UpdateTaskStatusRequest request = new UpdateTaskStatusRequest("done");
-        TaskResponse response = new TaskResponse(10L, "Task A", "Description", LocalDateTime.of(2026, 1, 10, 0, 0), "done", 1L, 2L, 100L, "101");
-        when(taskService.updateTaskStatus(10L, request, "auth-1")).thenReturn(response);
+        TaskResponse response = new TaskResponse((int) 10L, "Task A", "Description", LocalDateTime.of(2026, 1, 10, 0, 0), "done", (int) 1L, (int) 2L, (int) 100L, "101");
+        when(taskService.updateTaskStatus((int) 10L, request, "auth-1")).thenReturn(response);
 
-        ResponseEntity<TaskResponse> result = taskController.updateTaskStatus(10L, request, "auth-1");
+        ResponseEntity<TaskResponse> result = taskController.updateTaskStatus((int) 10L, request, "auth-1");
 
         assertEquals(HttpStatus.OK, result.getStatusCode());
         assertEquals(response, result.getBody());
@@ -56,9 +56,9 @@ class TaskControllerTest {
 
     @Test
     void deleteTask_returnsNoContent() {
-        doNothing().when(taskService).deleteTask(10L, "auth-1");
+        doNothing().when(taskService).deleteTask((int) 10L, "auth-1");
 
-        ResponseEntity<Void> result = taskController.deleteTask(10L, "auth-1");
+        ResponseEntity<Void> result = taskController.deleteTask((int) 10L, "auth-1");
 
         assertEquals(HttpStatus.NO_CONTENT, result.getStatusCode());
         assertEquals(null, result.getBody());
